requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
